requirements.txt
Flask==2.3.3
tensorflow==2.19.0
mediapipe==0.10.21
numpy==1.26.4
protobuf==4.25.3
opencv-python==4.9.0.80
Pillow==10.2.0
SignLanguageRecognition
music21==9.5.0
vexflow==5.0.0